simplify project structure
