+++ lib.rs
@@ -1,5 +1,5 @@
 #![cfg_attr(not(feature = "std"), no_std, no_main)]
 
 pub mod psp34;
-pub mod traits;
 pub mod types;
+pub mod traits;
