--- psp34/mod.rs
@@ -1,1 +1,0 @@
-pub mod psp34;
