--- traits/mod.rs
@@ -1,8 +1,0 @@
-pub mod psp34;
-
-pub use psp34::*;
-pub mod extensions {
-    pub mod psp34_enumerable;
-    pub mod psp34_metadata;
-    pub mod psp34_mintable;
-}
